Enable list actions
The baseRequest has been split in two functions, one for the TMDB,
the other for the AWS requests.

Button to add film to watched has been commented out with all of
its logic, as it is to be redesigned.

Typing extended to accommodate for all of the new endpoints.

The logic of actions in the account module has been made reusable
for both addition and removal of a film.